+++ gamemode/cl_init.lua
@@ -1,2 +1,4 @@
 -- This file is for client-side initialization
 include( "shared.lua" )
+
+usermessage.Hook( "TeamMenu", ChooseTeam )